LPS-137148 Move logic to update an item's common style to separated function, so we can reuse it somewhere else
